+++ pallas-applying/tests/alonzo.rs
@@ -4,7 +4,9 @@ use common::*;
 
 use pallas_addresses::{Address, Network, ShelleyAddress, ShelleyPaymentPart};
 use pallas_applying::{
-    utils::{AlonzoError, Environment, ValidationError::*},
+    utils::{
+        AlonzoError, AlonzoProtParams, Environment, MultiEraProtocolParameters, ValidationError::*,
+    },
     validate, UTxOs,
 };
 use pallas_codec::{
@@ -15,11 +17,11 @@ use pallas_codec::{
     utils::{Bytes, KeepRaw, KeyValuePairs, Nullable},
 };
 use pallas_primitives::alonzo::{
-    AddrKeyhash, AlonzoProtParams, CostModel, ExUnitPrices, ExUnits, Language, MintedTx,
-    MintedWitnessSet, NativeScript, NetworkId, Nonce, NonceVariant, PlutusData, RationalNumber,
-    Redeemer, RedeemerTag, TransactionBody, TransactionOutput, VKeyWitness, Value,
+    AddrKeyhash, CostModel, ExUnitPrices, ExUnits, Language, MintedTx, MintedWitnessSet,
+    NativeScript, NetworkId, Nonce, NonceVariant, PlutusData, RationalNumber, Redeemer,
+    RedeemerTag, TransactionBody, TransactionOutput, VKeyWitness, Value,
 };
-use pallas_traverse::{Era, MultiEraInput, MultiEraOutput, MultiEraProtocolParameters, MultiEraTx};
+use pallas_traverse::{Era, MultiEraInput, MultiEraOutput, MultiEraTx};
 use std::borrow::Cow;
 
 #[cfg(test)]
