+++ pallas-primitives/src/alonzo/model.rs
@@ -711,132 +711,6 @@ pub struct Update {
     pub epoch: Epoch,
 }
 
-#[derive(Debug, Clone, Encode, Decode)]
-pub struct ShelleyProtParams {
-    #[n(0)]
-    pub minfee_a: u32,
-
-    #[n(1)]
-    pub minfee_b: u32,
-
-    #[n(2)]
-    pub max_block_body_size: u32,
-
-    #[n(3)]
-    pub max_transaction_size: u32,
-
-    #[n(4)]
-    pub max_block_header_size: u32,
-
-    #[n(5)]
-    pub key_deposit: Coin,
-
-    #[n(6)]
-    pub pool_deposit: Coin,
-
-    #[n(7)]
-    pub maximum_epoch: Epoch,
-
-    #[n(8)]
-    pub desired_number_of_stake_pools: u32,
-
-    #[n(9)]
-    pub pool_pledge_influence: RationalNumber,
-
-    #[n(10)]
-    pub expansion_rate: UnitInterval,
-
-    #[n(11)]
-    pub treasury_growth_rate: UnitInterval,
-
-    #[n(12)]
-    pub decentralization_constant: UnitInterval,
-
-    #[n(13)]
-    pub extra_entropy: Nonce,
-
-    #[n(14)]
-    pub protocol_version: ProtocolVersion,
-
-    #[n(15)]
-    pub min_utxo_value: Coin,
-}
-
-#[derive(Debug, Clone, Encode, Decode)]
-pub struct AlonzoProtParams {
-    #[n(0)]
-    pub minfee_a: u32,
-
-    #[n(1)]
-    pub minfee_b: u32,
-
-    #[n(2)]
-    pub max_block_body_size: u32,
-
-    #[n(3)]
-    pub max_transaction_size: u32,
-
-    #[n(4)]
-    pub max_block_header_size: u32,
-
-    #[n(5)]
-    pub key_deposit: Coin,
-
-    #[n(6)]
-    pub pool_deposit: Coin,
-
-    #[n(7)]
-    pub maximum_epoch: Epoch,
-
-    #[n(8)]
-    pub desired_number_of_stake_pools: u32,
-
-    #[n(9)]
-    pub pool_pledge_influence: RationalNumber,
-
-    #[n(10)]
-    pub expansion_rate: UnitInterval,
-
-    #[n(11)]
-    pub treasury_growth_rate: UnitInterval,
-
-    #[n(12)]
-    pub decentralization_constant: UnitInterval,
-
-    #[n(13)]
-    pub extra_entropy: Nonce,
-
-    #[n(14)]
-    pub protocol_version: ProtocolVersion,
-
-    #[n(15)]
-    pub min_pool_cost: Coin,
-
-    #[n(16)]
-    pub ada_per_utxo_byte: Coin,
-
-    #[n(17)]
-    pub cost_models_for_script_languages: CostMdls,
-
-    #[n(18)]
-    pub execution_costs: ExUnitPrices,
-
-    #[n(19)]
-    pub max_tx_ex_units: ExUnits,
-
-    #[n(20)]
-    pub max_block_ex_units: ExUnits,
-
-    #[n(21)]
-    pub max_value_size: u32,
-
-    #[n(22)]
-    pub collateral_percentage: u32,
-
-    #[n(24)]
-    pub max_collateral_inputs: u32,
-}
-
 // Can't derive encode for TransactionBody because it seems to require a very
 // particular order for each key in the map
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
