+++ pallas-primitives/src/lib.rs
@@ -8,14 +8,3 @@ pub mod byron;
 pub mod conway;
 
 pub use framework::*;
-
-use pallas_codec::minicbor::{self, Decode, Encode};
-
-#[derive(Debug, Clone, Encode, Decode)]
-pub struct FeePolicy {
-    #[n(0)]
-    pub summand: u64,
-
-    #[n(1)]
-    pub multiplier: u64,
-}
